pyproject.toml
[project]
name = "lfimaging"
version = "0.1.0"
description = "LaminarForge ML imaging pipeline (Cellpose + U-Net) for organ-chip metrics"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [{ name = "LaminarForge Nonprofit" }]

dependencies = [
  "numpy>=1.26",
  "scipy>=1.12",
  "scikit-image>=0.22",
  "opencv-python-headless>=4.9",
  "tifffile>=2024.1.30",
  "imageio>=2.34",
  "paho-mqtt>=2.1",
  "boto3>=1.34",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "click>=8.1",
  "watchdog>=4.0",
  # ML stack — kept optional-by-default via extras because they're heavy.
  # Install with `pip install -e .[ml]` to pull in torch + cellpose.
]

[project.optional-dependencies]
ml = [
  "torch>=2.2",
  "torchvision>=0.17",
  "cellpose>=3.1.0",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "mypy>=1.10",
  "types-PyYAML",
]

[project.scripts]
lf-imaging = "lfimaging.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lfimaging"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
files = ["src/lfimaging"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
